refactor: extract utility methods to `fsUtils`
