+++ __tests__/unit/lib/utils/fsHelper.test.ts
@@ -6,8 +6,6 @@ import {
   pathExists,
   readDir,
   readPathFromGit,
-  sanitizePath,
-  treatPathSep,
   writeFile,
 } from '../../../../src/utils/fsHelper'
 import {
@@ -18,7 +16,6 @@ import { outputFile } from 'fs-extra'
 import { Work } from '../../../../src/types/work'
 import { Config } from '../../../../src/types/config'
 import { Ignore } from 'ignore'
-import { sep } from 'path'
 
 jest.mock('fs-extra')
 
@@ -391,64 +388,5 @@ describe('copyFile', () => {
       // Assert
       expect(outputFile).not.toBeCalled()
     })
-  })
-})
-
-describe('treatPathSep', () => {
-  it(`replace / by ${sep}`, () => {
-    // Arrange
-    const input = 'test///test//test/test'
-
-    // Act
-    const result = treatPathSep(input)
-
-    // Assert
-    expect(result).toBe(`test${sep}test${sep}test${sep}test`)
-  })
-
-  it(`replace \\ by ${sep}`, () => {
-    // Arrange
-    const input = 'test\\\\\\test\\\\test\\test'
-
-    // Act
-    const result = treatPathSep(input)
-
-    // Assert
-    expect(result).toBe(`test${sep}test${sep}test${sep}test`)
-  })
-})
-
-describe('sanitizePath', () => {
-  it(`returns path with '${sep}' separator`, () => {
-    // Arrange
-    const input = 'test\\test/test'
-
-    // Act
-    const result = sanitizePath(input)
-
-    // Assert
-    expect(result).toBe(`test${sep}test${sep}test`)
-  })
-
-  it(`normalize path`, () => {
-    // Arrange
-    const input = 'test/test\\../test'
-
-    // Act
-    const result = sanitizePath(input)
-
-    // Assert
-    expect(result).toBe(`test${sep}test`)
-  })
-
-  it('return empty string when data is empty string', () => {
-    // Arrange
-    const input = ''
-
-    // Act
-    const result = sanitizePath(input)
-
-    // Assert
-    expect(result).toBe('')
   })
 })
