+++ src/utils/cliHelper.ts
@@ -7,8 +7,7 @@ import {
   isVersionSupported,
 } from '../metadata/metadataManager'
 import { format } from 'util'
-import { sanitizePath } from './fsHelper'
-import { readFile, dirExists, fileExists } from './fsUtils'
+import { readFile, dirExists, fileExists, sanitizePath } from './fsUtils'
 import { join } from 'path'
 import { Work } from '../types/work'
 import { Config } from '../types/config'
