+++ src/utils/fsUtils.ts
@@ -1,7 +1,15 @@
 'use strict'
 import { stat, readFile as fsReadFile } from 'fs-extra'
-import { isAbsolute, relative } from 'path'
+import { isAbsolute, normalize, relative, sep } from 'path'
 import { UTF8_ENCODING } from '../constant/fsConstants'
+
+export const treatPathSep = (data: string) => data?.replace(/[/\\]+/g, sep)
+export const sanitizePath = (data: string) => {
+  if (data) {
+    return normalize(treatPathSep(data))
+  }
+  return data
+}
 
 export const isSubDir = (parent: string, dir: string) => {
   const rel = relative(parent, dir)
